refactor(app): extract useCampaignFlow (#316 phase 9c)
The last and largest handler group: all 25 campaign-run handlers, ~1,090
lines, moved verbatim. Launching an act, selecting a node, resolving every
node type, winning or losing a battle, completing an act, and abandoning or
retrying a run — including the deeply nested async chains in launchCampaign
and handleActComplete (relic spin -> relic select -> card rest -> next act).

Dependencies are one named object rather than positional arguments. The set is
68 wide and many entries are structurally identical — a dozen
Dispatch<SetStateAction<boolean>> among them — so positional passing would let
a swap typecheck silently. With shorthand property syntax at the call site a
swap means deliberately writing the wrong name against the right key.

That typing caught a real mistake while wiring: merchantBoughtRef was assumed
to be a flag and is actually a counter.

handleMainMenu moves above the hook call, since the campaign flow takes it as
a dependency. This also clears out the imports that were only feeding these
handlers, App.tsx's import block drops from ~120 lines to ~55.

App.tsx: 2,321 -> 1,261 lines. Build green, chunk boundaries unchanged,
2107 tests passing.

Verified in the browser as well, since typecheck cannot catch a mis-wired
callback: CAMPAIGN launches, the async act load resolves into the four-panel
intro cutscene, and Enter lands on the Act I node map with live run state
(HP 50/50, three lives, consumables). No application errors. Node selection
itself goes through PixiJS canvas pathfinding, which AGENTS.md calls out as
expensive to drive, so it was left to the same handleSelectNode reference
CampaignRoutes already passes to NodeMap.
